requirements.txt
python-pptx
canvasapi
browser_cookie3